fix arm64 gmon.out. fix lr and last lr.

+++ jni/read_maps.h
@@ -11,14 +11,14 @@
 #define read_maps_h_seen
 
 struct proc_map {
-	unsigned int base;
-	unsigned int lo;
-	unsigned int hi;
+	unsigned long base;
+	unsigned long lo;
+	unsigned long hi;
 	struct proc_map *next;
 };
 
 struct proc_map *read_maps(FILE *fp, const char *lname);
 void free_maps(struct proc_map *s);
-unsigned int get_real_address(const struct proc_map *maps, unsigned int fake);
+unsigned long get_real_address(const struct proc_map *maps, unsigned long fake);
 
 #endif
